requirements.txt
backports.zoneinfo    0.2.1
certifi               2022.12.7
charset-normalizer    3.0.1
clickhouse-connect    0.5.3
DateTime              5.0
idna                  3.4
iso8601               1.1.0
lz4                   4.3.2
numpy                 1.24.1
pandas                1.5.3
pip                   22.3.1
python-dateutil       2.8.2
pytz                  2022.7.1
pytz-deprecation-shim 0.1.0.post0
requests              2.28.2
sanpy                 0.11.3
setuptools            66.1.1
six                   1.15.0
tzdata                2022.7
tzlocal               4.2
urllib3               1.26.14
wheel                 0.33.1
zope.interface        5.5.2
zstandard             0.19.0
